reduces to one component and cleans styling
