--- kahuna/public/js/components/gr-add-preset-label/gr-add-preset-label.css
@@ -1,31 +1,0 @@
-gr-add-preset-label{
-    display: flex;
-    height: 25px;
-    padding: 5px;
-}
-
-gr-add-preset-label:hover {
-    color: white;
-}
-
-gr-add-preset-label .small {
-    border: none;
-    padding: 0;
-}
-
-.gr-add-preset-label__form {
-    display: flex;
-}
-
-.gr-add-preset-label__form__input {
-    width: 150px;
-}
-
-.gr-add-preset-label__form__buttons {
-    display: flex;
-    font-size: 13px;
-}
-
-.gr-add-preset-label__form__buttons__button-cancel, .gr-add-preset-label__form__buttons__button-save{
-    margin-left: 5px;
-}
