--- kahuna/public/js/upload/prompt/preset-labeller.html
@@ -1,16 +1,0 @@
-<div class="preset-labeller">
-    <ul class="preset-labels"
-        ng:if="ctrl.presetLabels.length > 0">
-        <li class="preset-label"
-            ng:repeat="label in ctrl.presetLabels">
-            <a ui:sref="search.results({query: (label | queryLabelFilter)})"
-               class="preset-label__value preset-label__link">{{label}}</a>
-
-            <button class="preset-label__remove"
-                    title="Remove label"
-                    ng:click="ctrl.removePresetLabel(label)">
-                <gr-icon>close</gr-icon>
-            </button>
-        </li>
-    </ul>
-</div>
