+++ kahuna/public/js/upload/prompt/prompt.html
@@ -3,7 +3,9 @@
 </div>
 
 <div class="preset-labels">
-    <span class="preset-labels__heading">Default labels for uploaded images: </span>
-    <ui-preset-labeller></ui-preset-labeller>
-    <gr-add-preset-label></gr-add-preset-label>
+    <p class="preset-labels__heading">Apply
+        <span ng:if="ctrl.presetLabels.length === 0 && !ctrl.active">label e.g. Observer </span>
+    </p>
+    <gr-preset-labels></gr-preset-labels>
+    <p class="preset-labels__heading">to all uploads</p>
 </div>
